pyproject.toml
[tool.poetry]
name = "solarmanagement"
version = "0.1.0"
description = ""
authors = ["andreas weier <andreas.weier.open.source@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pymodbus = "^3.5.4"
pyyaml = "^5.3.1"
ShellyPy = "^0.3.0"
solaredge_modbus = "^0.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

